+++ application/views/main-frame.php
@@ -7,6 +7,15 @@
   <?php $this->load->view('includes/head'); ?>
   <body>
     <!--[if lt IE 7]><script>for (x in open);</script><![endif]-->
+    <div id="fb-root"></div>
+    <script>(function(d, s, id) {
+      var js, fjs = d.getElementsByTagName(s)[0];
+      if (d.getElementById(id)) return;
+      js = d.createElement(s); js.id = id;
+      js.src = "//connect.facebook.net/en_US/sdk.js#xfbml=1&version=v2.3";
+      fjs.parentNode.insertBefore(js, fjs);
+    }(document, 'script', 'facebook-jssdk'));</script>
+    <!--load page-->
     <?php $this->load->view('includes/main-header'); ?>
     <div id="content" class="pure-g-r" data-page="<?php echo $page; ?>">
       <?php $this->load->view('main/'.$page); ?>